refactor(portal): extract server env into its own file for safety
