+++ apps/mini-booking-portal/src/constants/server-api.ts
@@ -1,0 +1,4 @@
+import 'server-only';
+
+export const BOOKING_API_URL =
+  process.env.BOOKING_API_URL ?? 'http://localhost:8080/api';
